requirements.txt
aiofiles==23.2.1
aiogram==2.23.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.6
Jinja2==3.1.3
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
turtle==0.0.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.4
